Grouped asset types database methods in their own file
